New feature: transference between accounts.
Create the 'transference ' context to handle the transfer operation and
all the validations and sub-operations involving it.

Create the 'transference' controller, and add its operations to the
router at the right scope.

Change the 'account' schema to handle a database update operation
at the field 'balance'.

Create all the needed tests.