New feature: withdraw money from an account.
Create the 'withdraw' context to handle the withdraw operation and all
the validations and sub-operations involving it.

Create the 'withdraw' controller, and add its operations to the router
at the right scope.

Create all the needed tests.

Fix 'transference_test' to remove used data on fixtures.

Change 'bcrypt' configuration to run tests faster.